fixes for code climate
