+++ src/admin/components/website/modalParts/emailOverride/WebsiteFeedbackEmailOverride.tsx
@@ -18,9 +18,10 @@
 
 import * as React from 'react';
 import { connect } from 'react-redux';
-import {WebsiteOverrideSettings} from 'admin/components/website/modalParts/WebsiteOverrideSettings';
+import {WebsiteOverrideSettings} from './WebsiteOverrideSettings';
+import { WebsiteBaseOverride } from './WebsiteBaseOverride';
 
-class WebsiteEscalationEmailOverride extends React.Component <any, any> {
+class WebsiteFeedbackEmailOverride extends WebsiteBaseOverride {
 
 	constructor (props) {
 		super(props);
@@ -28,42 +29,28 @@ class WebsiteEscalationEmailOverride extends React.Component <any, any> {
 			value: '',
 			touched: false,
 		};
-
 		this.onChange = this.onChange.bind(this);
 		this.sendOverrideChanges = this.sendOverrideChanges.bind(this);
 	}
 
-	onChange(value:string, touched:boolean) {
-		this.setState({
-			value,
-			touched,
-		});
-	}
-
-	sendOverrideChanges (data) {
-		const {overrideSettings: os} = this.props;
-		const overrideSettings = Object.assign({}, os, data);
-		this.setState({value: '', touched: false});
-		return this.props.onChange({overrideSettings});
-	}
-
 	render () {
-		const {settings: {escalation: escalationStatus = false}, escalationEmailOverride} = this.props;
+		const {settings: {feedback: feedbackOverrideStatus = false}, feedbackEmailOverride} = this.props;
 		const {value, touched} = this.state;
 		return (
 			<WebsiteOverrideSettings
-				list={escalationEmailOverride}
-				listPropName={'escalationEmail'}
+				list={feedbackEmailOverride}
+				listPropName={'feedbackEmail'}
 				isControllable={true}
-				controlPropName={'escalation'}
-				controlPropValue={escalationStatus}
-				label={<span>
-					<b>Escalation email override</b><br/>
-					Emails to send notifications about escalation (instead of website chief editors)
-				</span>}
-				ID={'escalation-email-override-status'}
+				controlPropName={'feedback'}
+				controlPropValue={feedbackOverrideStatus}
 				value={value}
 				touched={touched}
+				label={<span>
+					<b>Feedback email override</b><br/>
+					Emails to get notifications about feedbacks for articles if needed
+					(instead of article authors emails)
+				</span>}
+				ID={'feedback-email-override-status'}
 				onSubmit={this.sendOverrideChanges}
 				onChange={this.onChange}
 			/>
@@ -73,8 +60,8 @@ class WebsiteEscalationEmailOverride extends React.Component <any, any> {
 
 const mapStateToProps = (state, ownProps) => {
 	return {
-		escalationEmailOverride: state.website.getIn(
-			['selected', 'overrideSettings', 'overrides', 'escalationEmail'], []),
+		feedbackEmailOverride: state.website.getIn(
+			['selected', 'overrideSettings', 'overrides', 'feedbackEmail'], []),
 		overrideSettings: state.website.getIn(
 			['selected', 'overrideSettings'], {}),
 		settings: state.website.getIn(
@@ -86,4 +73,4 @@ const mapDispatchToProps = (dispatch, ownProps) => {
 	return {};
 };
 
-export default connect(mapStateToProps, mapDispatchToProps)(WebsiteEscalationEmailOverride);
+export default connect(mapStateToProps, mapDispatchToProps)(WebsiteFeedbackEmailOverride);
